fix: Move client to commons/ folder
